+++ adminforth/spa/src/composables/useFrontendApi.ts
@@ -1,15 +1,17 @@
+import adminforth from '@/adminforth';
+
 export function showSuccesTost(message: string) {
-  window.adminforth.alert({ message, variant: 'success' });  
+  adminforth.alert({ message, variant: 'success' });  
   return message;
 }
 
 export function showWarningTost(message: string) {
-  window.adminforth.alert({ message, variant: 'warning' });  
+  adminforth.alert({ message, variant: 'warning' });  
   return message;
 }
 
 export function showErrorTost(message: string, timeout?: number) {
-  window.adminforth.alert({ message, variant: 'danger', timeout: timeout || 30});  
+  adminforth.alert({ message, variant: 'danger', timeout: timeout || 30});  
   return message;
 }